extract storage setup from startup script.  disable PM2 in docker
