+++ checkapi/config.js
@@ -3,5 +3,5 @@ module.exports = {
   MONGO_CONNECTION_STRING: process.env.MONGO_CONNECTION_STRING || 'mongodb://localhost/mtc',
   PORT: process.env.PORT || '3003',
   APPINSIGHTS_INSTRUMENTATIONKEY: process.env.APPINSIGHTS_INSTRUMENTATIONKEY,
-  CheckStorage: 'TableStorage' // 'Mongoose' 'MongoOfficial'
+  StorageProvider: process.env.STORAGE_PROVIDER
 }
