+++ checkapi/package.json
@@ -7,10 +7,11 @@
   "author": "Guy Harwood",
   "license": "GPL-3.0",
   "scripts": {
-    "start": "npm run serve",
+    "start": "node ./index.js",
     "build": "standard",
-    "serve": "nodemon ./index.js",
-    "test": "nyc -r html ./run-jasmine.js"
+    "debug": "nodemon ./index.js",
+    "test": "nyc -r html ./run-jasmine.js",
+    "pm2": "pm2 start ./index.js"
   },
   "dependencies": {
     "applicationinsights": "^1.0.0",
